+++ apps/geography/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
+from apps.core.localization import localized_value
 from apps.core.mixins import ActiveMixin, LocalizedNameMixin, LocalizedSlugMixin
 from apps.core.models import BaseModel
 
@@ -33,6 +34,26 @@ class Province(BaseModel, LocalizedNameMixin, LocalizedSlugMixin, ActiveMixin):
 
 class City(BaseModel, LocalizedNameMixin, LocalizedSlugMixin, ActiveMixin):
     province = models.ForeignKey(Province, on_delete=models.PROTECT, related_name="cities")
+    description_en = models.TextField(blank=True)
+    description_fa = models.TextField(blank=True)
+    description_tr = models.TextField(blank=True)
+    description_ar = models.TextField(blank=True)
+    seo_title_en = models.CharField(max_length=255, blank=True)
+    seo_title_fa = models.CharField(max_length=255, blank=True)
+    seo_title_tr = models.CharField(max_length=255, blank=True)
+    seo_title_ar = models.CharField(max_length=255, blank=True)
+    seo_description_en = models.TextField(blank=True)
+    seo_description_fa = models.TextField(blank=True)
+    seo_description_tr = models.TextField(blank=True)
+    seo_description_ar = models.TextField(blank=True)
+
+    @property
+    def localized_seo_title(self):
+        return localized_value(self, "seo_title")
+
+    @property
+    def localized_seo_description(self):
+        return localized_value(self, "seo_description")
 
     def __str__(self):
         return self.localized_name